+++ includes/get_next_line.h
@@ -23,7 +23,7 @@ char	*get_next_line(int fd);
 char	*ft_read_to_static(int fd, char *left_str);
 char	*ft_strchr(char *s, int c);
 char	*ft_strjoin(char *left_str, char *buff);
-size_t	ft_strlen(char *s);
+int 	ft_strlen(char *s);
 char	*ft_get_line(char *left_str);
 char	*ft_new_left_str(char *left_str);
